Perf: hot-path round 1 — tick SIMD, seam detach, u16 hist (768→384 MiB), bake caching, troubleshoot logs
docs/PERF-MAP.md: the thermal-budget re-map (tick/burst/post rings, hazard audit,
iOS 27 adoption points, S/K/I signal section) + §6 ledger of everything below.

H1  KernelsPalette16: hlg10 range pre-scan -> SIMD16 lane-max (same refuse-before-
    write contract); ColorHead x420 loop: chroma offsets once per 4:2:0 pixel pair.
H3  poolV21Counts (~201M adds) off the shutter seam into the detached flow task;
    new v21CountsCallback (epoch-guarded); BurstResult.v21Counts nil by design.
H2  v21 hist buffer counts i32 -> u16 = 384 MiB (cells thread-owned, no atomics);
    readers widen at read; scale>63 guard + one-shot log; parity tests re-pinned.
H4  InvertedPyramidField: pool+bake per pixel-change (not per body eval); shutter
    rebakes 256 cells only when the filled-cell count steps.
H7  s4_set_log_gate: suppression checked BEFORE formatting (KernelsCore + ABI).
H6/H10  MetalPipeline per-frame texture pool (recycle after readback, cap 4);
    BandHeadTrainer.shared replaces per-burst GPUContext+PSO compile.
ColorHead t-band pairs: [[Int64]] -> flat stride-4 (no per-block heap arrays).
LUT export DEPRECATED behind Feature.lutExport=false (kernels stay golden-gated).
Batch 0: app-target excludes for 8 mechanically-verified zero-ref files.
Troubleshoot logs (aggregate, never per-tick): [perf] yin-yang tick CPU per burst,
[tick] LATE frame anomaly-only, [perf] texture pool miss #N, [perf] hist MiB alloc.

Gates: full XCTest+Swift Testing suite, GRID lint, BUILD SUCCEEDED (iPhone 17 Pro sim).

+++ SixFour/Metal/Shaders.metal
@@ -266,10 +266,16 @@ kernel void cropDownsampleLinearizeKernel(
 // (the same bounds discipline as `cropDownsampleLinearizeKernel`), which only
 // arises at the crop margin; the Zig assumes divisible, fully-in-bounds boxes.
 
+// PERF 2026-07-08: counts are u16, not i32 — this buffer is the app's largest
+// allocation (64 frames × 64² × 3 × 256 levels; i32 made it 768 MiB, u16 is
+// 384 MiB). Capacity: a (voxel,channel,level) cell holds at most scale² (hard)
+// or scale²·wBudget (soft) — ≤ 65535 for every scale ≤ 63 (crop min-dim ≤ ~4K;
+// the host guards and skips the dispatch above that). No atomics needed (one
+// thread owns its cells), so the narrower type is free.
 kernel void v21AccumulateHistKernel(
     texture2d<float, access::read> luma   [[texture(0)]],
     texture2d<float, access::read> chroma [[texture(1)]],
-    device int*    outCounts        [[buffer(0)]],
+    device ushort* outCounts        [[buffer(0)]],
     constant int2  &srcOffset       [[buffer(1)]],
     constant int   &scale           [[buffer(2)]],
     constant uchar &colorSpaceTag   [[buffer(3)]],
@@ -347,7 +353,7 @@ kernel void v21AccumulateHistKernel(
 kernel void v21AccumulateHistSoftKernel(
     texture2d<float, access::read> luma   [[texture(0)]],
     texture2d<float, access::read> chroma [[texture(1)]],
-    device int*    outCounts        [[buffer(0)]],
+    device ushort* outCounts        [[buffer(0)]],   // u16 counts — see the capacity note above
     constant int2  &srcOffset       [[buffer(1)]],
     constant int   &scale           [[buffer(2)]],
     constant uchar &colorSpaceTag   [[buffer(3)]],
@@ -403,9 +409,9 @@ kernel void v21AccumulateHistSoftKernel(
                 int vlo  = hi[ch] / wBudget;
                 int frac = hi[ch] % wBudget;
                 int chBase = cellBase + ch * nLevels;
-                outCounts[chBase + vlo] += (wBudget - frac);
+                outCounts[chBase + vlo] += (ushort)(wBudget - frac);
                 if (frac != 0 && vlo + 1 < nLevels) {
-                    outCounts[chBase + vlo + 1] += frac;
+                    outCounts[chBase + vlo + 1] += (ushort)frac;
                 }
             }
         }
